requirements.txt
natsort
frozendict
absl-py
transformers[torch]
datasets
fuzzywuzzy
thefuzz[speedup]